Updated: code to use the same component in the open and packed section and used slots to update the content(#1w40vqa)
